+++ packages/grida-cmath/_snap.ts
@@ -1,6 +1,31 @@
 import assert from "assert";
 import { cmath } from ".";
 
+interface IDistance {
+  distance: number;
+}
+
+function bestAxisAlignedDistance(...results: IDistance[]): IDistance {
+  let min_distance = Infinity;
+  let min_index = -1;
+
+  for (let i = 0; i < results.length; i++) {
+    const result = results[i];
+    if (Math.abs(result.distance) < Math.abs(min_distance)) {
+      min_distance = result.distance;
+      min_index = i;
+    }
+    // TODO: support multiple results, merge them if the distance is the same (or within tolerance)
+    // else if (result.distance === min_distance) { }
+  }
+
+  if (min_index === -1) {
+    return results[0];
+  }
+
+  return results[min_index];
+}
+
 export type SnapToObjectsResult = {
   /**
    * The translated rectangle after snapping.
@@ -19,12 +44,15 @@ export type SnapToObjectsResult = {
    */
   delta: cmath.Vector2;
 
+  // TODO: group by x/y, allow multiple results by type (if matches)
+  // x: SnapToObjectsAxisResult [{ type: 'by_geometry': ... }, ...]
+
   by_geometry: {
     x: cmath.ext.snap.Snap1DResult | null;
     y: cmath.ext.snap.Snap1DResult | null;
     hit_points: {
-      agent: ObjectGeometryHitResult;
-      anchors: ObjectGeometryHitResult[];
+      agent: Object9PointGeometryHitResult;
+      anchors: Object9PointGeometryHitResult[];
     };
   };
 
@@ -36,6 +64,7 @@ export type SnapToObjectsResult = {
       | (Snap1DRangesDirectionAlignedResult & { aligned_anchors_idx: number[] })
       | null;
   };
+
   by_ruler: {};
 };
 
@@ -48,7 +77,12 @@ export function snapToObjects(
   assert(agent, "Agent must be a valid rectangle.");
   assert(anchors.length > 0, "Anchors must contain at least one rectangle.");
 
-  const snap_geo = snapToObjectsGeometry(agent, anchors, threshold, tolerance);
+  const snap_geo = snapToObjects9PointsGeometry(
+    agent,
+    anchors,
+    threshold,
+    tolerance
+  );
   const snap_spc = snapToObjectsSpace(agent, anchors, threshold, tolerance);
   const x = bestAxisAlignedDistance(snap_geo.x, snap_spc.x);
   const y = bestAxisAlignedDistance(snap_geo.y, snap_spc.y);
@@ -79,36 +113,11 @@ export function snapToObjects(
   };
 }
 
-interface IDistance {
-  distance: number;
-}
-
-function bestAxisAlignedDistance(...results: IDistance[]): IDistance {
-  let min_distance = Infinity;
-  let min_index = -1;
-
-  for (let i = 0; i < results.length; i++) {
-    const result = results[i];
-    if (Math.abs(result.distance) < Math.abs(min_distance)) {
-      min_distance = result.distance;
-      min_index = i;
-    }
-    // TODO: support multiple results, merge them if the distance is the same (or within tolerance)
-    // else if (result.distance === min_distance) { }
-  }
-
-  if (min_index === -1) {
-    return results[0];
-  }
-
-  return results[min_index];
-}
-
-type ObjectGeometryHitResult = cmath.rect.TRectangle9PointsChunk<
+type Object9PointGeometryHitResult = cmath.rect.TRectangle9PointsChunk<
   [boolean, boolean]
 >;
 
-function snapToObjectsGeometry(
+function snapToObjects9PointsGeometry(
   agent: cmath.Rectangle,
   anchors: cmath.Rectangle[],
   threshold: cmath.Vector2,
@@ -116,8 +125,8 @@ function snapToObjectsGeometry(
 ): {
   x: cmath.ext.snap.Snap1DResult;
   y: cmath.ext.snap.Snap1DResult;
-  agent_hits: ObjectGeometryHitResult;
-  anchor_hits: ObjectGeometryHitResult[];
+  agent_hits: Object9PointGeometryHitResult;
+  anchor_hits: Object9PointGeometryHitResult[];
   agent_points: cmath.Vector2[];
   anchor_points: cmath.ext.snap.AxisAlignedPoint[];
 } {
@@ -128,22 +137,22 @@ function snapToObjectsGeometry(
     .map((r) => cmath.rect.to9PointsChunk(r))
     .flat();
 
-  const snap = snap2DAxisAligned(
+  const snap = cmath.ext.snap.snap2DAxisAligned(
     agent_points,
     anchor_points,
     threshold,
     tolerance
   );
 
-  const agent_hits: ObjectGeometryHitResult = agent_points.map(
+  const agent_hits: Object9PointGeometryHitResult = agent_points.map(
     (point, index) => {
       const xHit = snap.x.hit_agent_indices.includes(index);
       const yHit = snap.y.hit_agent_indices.includes(index);
       return [xHit, yHit] satisfies [boolean, boolean];
     }
-  ) as ObjectGeometryHitResult;
+  ) as Object9PointGeometryHitResult;
 
-  const anchor_hits: ObjectGeometryHitResult[] = [];
+  const anchor_hits: Object9PointGeometryHitResult[] = [];
   for (let i = 0; i < anchor_points.length; i += geometry_chunk_size) {
     const chunk = anchor_points.slice(i, i + geometry_chunk_size);
     const hitResult = chunk.map((point, index) => {
@@ -152,7 +161,7 @@ function snapToObjectsGeometry(
       const yHit = snap.y.hit_anchor_indices.includes(pointIndex);
       return [xHit, yHit];
     });
-    anchor_hits.push(hitResult as ObjectGeometryHitResult);
+    anchor_hits.push(hitResult as Object9PointGeometryHitResult);
   }
 
   return {
@@ -306,62 +315,322 @@ function snap1DRangesDirectionAlignedWithDistributionGeometry(
   };
 }
 
-type Sanp2DAxisAlignedResult = {
-  x: cmath.ext.snap.Snap1DResult;
-  y: cmath.ext.snap.Snap1DResult;
-};
+function __surface_snap_guide_by_geometry(context: SnapToObjectsResult) {
+  const { by_geometry, translated, anchors, delta } = context;
 
-/**
- * Snaps an array of points to the nearest target point along each axis independently.
- * The snapping delta is computed for each axis separately and applied to all points.
- *
- * @param agents - An array of 2D points (Vector2) to snap.
- * @param anchors - An array of existing 2D points to snap to.
- * @param threshold - The maximum allowed single-axis distance for snapping.
- * @returns The snapped points and the delta applied:
- *          - `value`: The translated points.
- *          - `distance`: The delta vector applied to align the points.
- */
-export function snap2DAxisAligned(
-  agents: cmath.Vector2[],
-  anchors: cmath.ext.snap.AxisAlignedPoint[],
-  threshold: cmath.Vector2,
-  tolerance = 0
-): Sanp2DAxisAlignedResult {
-  assert(agents.length > 0, "Agents must contain at least one point.");
-  assert(anchors.length > 0, "Anchors must contain at least one point.");
-  assert(threshold[0] >= 0, "Threshold must be a non-negative number.");
-  assert(threshold[1] >= 0, "Threshold must be a non-negative number.");
+  const { x, y } = by_geometry;
 
-  // Separate the scalar points for each axis
-  const x_agent_points = agents.map(([x]) => x);
-  const y_agent_points = agents.map(([_, y]) => y);
+  const lines: cmath.ui.Line[] = [];
+  const points: cmath.Vector2[] = [];
 
-  // Separate anchor points into x and y components
-  const x_anchor_points = anchors
-    .map(([x]) => x)
-    .filter((x): x is number => x !== null);
-  const y_anchor_points = anchors
-    .map(([_, y]) => y)
-    .filter((y): y is number => y !== null);
+  // Separate x-hit and y-hit points
+  const xPoints: cmath.Vector2[] = [];
+  const yPoints: cmath.Vector2[] = [];
 
-  // snap each axis
-  const x_snap = cmath.ext.snap.snap1D(
-    x_agent_points,
-    x_anchor_points,
-    threshold[0],
-    tolerance
+  by_geometry.hit_points.anchors.forEach((hit, i) => {
+    const anchor9 = cmath.rect.to9PointsChunk(anchors[i]);
+    hit.forEach(([xhit, yhit], j) => {
+      if (x && xhit) xPoints.push(anchor9[j]);
+      if (y && yhit) yPoints.push(anchor9[j]);
+      if ((x && xhit) || (y && yhit)) points.push(anchor9[j]);
+    });
+  });
+
+  const agent9 = cmath.rect.to9PointsChunk(translated);
+  by_geometry.hit_points.agent.forEach(([xhit, yhit], i) => {
+    if (x && xhit) xPoints.push(agent9[i]);
+    if (y && yhit) yPoints.push(agent9[i]);
+    if ((x && xhit) || (y && yhit)) points.push(agent9[i]);
+  });
+
+  // Vertical lines from xPoints
+  const xs = new Map<number, number[]>();
+  xPoints.forEach(([x, y]) => {
+    if (!xs.has(x)) xs.set(x, []);
+    xs.get(x)!.push(y);
+  });
+  xs.forEach((arrY, x) => {
+    if (arrY.length > 1) {
+      lines.push({
+        x1: x,
+        y1: Math.min(...arrY),
+        x2: x,
+        y2: Math.max(...arrY),
+      });
+    }
+  });
+
+  // Horizontal lines from yPoints
+  const ys = new Map<number, number[]>();
+  yPoints.forEach(([x, y]) => {
+    if (!ys.has(y)) ys.set(y, []);
+    ys.get(y)!.push(x);
+  });
+  ys.forEach((arrX, y) => {
+    if (arrX.length > 1) {
+      lines.push({
+        x1: Math.min(...arrX),
+        y1: y,
+        x2: Math.max(...arrX),
+        y2: y,
+      });
+    }
+  });
+
+  return { points, lines };
+}
+
+function __calc_spacing_loop_gap_line({
+  loop,
+  gap,
+  axis,
+}: {
+  loop: cmath.Rectangle[];
+  gap: number;
+  axis: cmath.Axis;
+}) {
+  const origianl_rect_first = loop[0];
+  const origianl_rect_last = loop[loop.length - 1];
+
+  const label = cmath.ui.formatNumber(gap, 1);
+
+  const counterAxis = cmath.counterAxis[axis];
+
+  const loop_gap_counter_axis_pos = cmath.range.mean(
+    cmath.range.fromRectangle(origianl_rect_first, counterAxis),
+    cmath.range.fromRectangle(origianl_rect_last, counterAxis)
+  );
+  // r.x + r.width
+  const loop_gap_main_axis_a = cmath.range.fromRectangle(
+    origianl_rect_first,
+    axis
+  )[1];
+  const loop_gap_main_axis_b = loop_gap_main_axis_a + gap;
+
+  const a = cmath.vector2.axisOriented(
+    loop_gap_main_axis_a,
+    loop_gap_counter_axis_pos,
+    axis
   );
 
-  const y_snap = cmath.ext.snap.snap1D(
-    y_agent_points,
-    y_anchor_points,
-    threshold[1],
-    tolerance
+  const b = cmath.vector2.axisOriented(
+    loop_gap_main_axis_b,
+    loop_gap_counter_axis_pos,
+    axis
   );
+
+  return cmath.ui.normalizeLine({
+    label: label,
+    x1: a[0],
+    y1: a[1],
+    x2: b[0],
+    y2: b[1],
+  } satisfies cmath.ui.Line);
+}
+
+function __calc_spacing_agent_gap_line({
+  p,
+  axis,
+  anchor,
+}: {
+  p: cmath.ext.snap.spacing.ProjectionPoint;
+  axis: cmath.Axis;
+  anchor: cmath.Rectangle;
+}) {
+  const lines: cmath.ui.Line[] = [];
+  const { p: pos, o: origin } = p;
+
+  // We'll pick a "counterAxis" coordinate (like the mid Y for axis="x", or mid X for axis="y")
+  const counterAxis = cmath.counterAxis[axis];
+  const anchorRectMid = cmath.range.mean(
+    cmath.range.fromRectangle(anchor, counterAxis)
+  );
+
+  // Convert anchor -> pos into a 2D line
+  // "anchor" is the point from which "pos" was derived,
+  // and they are both 1D along `axis`. So we pick anchorRectMid for the other coordinate
+  const anchorPt = cmath.vector2.axisOriented(origin, anchorRectMid, axis);
+  const posPt = cmath.vector2.axisOriented(pos, anchorRectMid, axis);
+  const gap = Math.abs(pos - origin);
+
+  const label = cmath.ui.formatNumber(gap, 1);
+
+  lines.push({
+    label: label,
+    x1: anchorPt[0],
+    y1: anchorPt[1],
+    x2: posPt[0],
+    y2: posPt[1],
+  });
+
+  return lines;
+}
+
+function __surface_snap_guide_by_spacing(context: SnapToObjectsResult) {
+  const { by_spacing, anchors: main_anchors } = context;
+
+  const { x, y } = by_spacing;
+  const lines: cmath.ui.Line[] = [];
+
+  function handle_axis({
+    a_flat, // flattened [pos, anchor]
+    a_flat_loops_idx, // flattened -> loop index
+    a_snap,
+    b_flat,
+    b_flat_loops_idx,
+    b_snap,
+    loops,
+    gaps,
+    aligned_anchors_idx,
+    anchors,
+    axis,
+    distance,
+  }: Snap1DRangesDirectionAlignedResult & {
+    aligned_anchors_idx: number[];
+    anchors: cmath.Rectangle[];
+    axis: cmath.Axis;
+  }) {
+    // If we actually snapped via the "a" side
+    if (a_snap.distance === distance) {
+      // Each anchor index we actually snapped to
+      a_snap.hit_anchor_indices.forEach((hitIdx) => {
+        const p = a_flat[hitIdx];
+        const { fwd } = p;
+        const loop_idx = a_flat_loops_idx[hitIdx];
+        const loop = loops[loop_idx];
+        const anchor_rect_idx =
+          aligned_anchors_idx[
+            loop[
+              // // fwd === -1 => center extension
+              fwd === -1 ? 0 : loop.length - 1
+            ]
+          ];
+        const anchor = anchors[anchor_rect_idx];
+
+        // the main line for the agent.
+        lines.push(...__calc_spacing_agent_gap_line({ p, axis, anchor }));
+
+        // lines for uniform gap loops (including self)
+
+        if (fwd !== -1) {
+          const loop = loops[fwd];
+          const gap = gaps[fwd];
+          const rects_loop = loop.map(
+            (idx) => anchors[aligned_anchors_idx[idx]]
+          );
+
+          lines.push(
+            __calc_spacing_loop_gap_line({
+              axis,
+              loop: rects_loop,
+              gap: gap,
+            })
+          );
+        }
+      });
+    }
+
+    // If we actually snapped via the "b" side
+    if (b_snap.distance === distance) {
+      // Each anchor index we actually snapped to
+      b_snap.hit_anchor_indices.forEach((hitIdx) => {
+        const p = b_flat[hitIdx];
+        const { fwd } = p;
+        const loop_idx = b_flat_loops_idx[hitIdx];
+        const loop = loops[loop_idx];
+        const anchor_rect_idx =
+          aligned_anchors_idx[
+            loop[
+              // fwd === -1 => center extension
+              fwd === -1 ? loop.length - 1 : 0
+            ]
+          ];
+        const anchor = anchors[anchor_rect_idx];
+
+        // the main line for the agent.
+        lines.push(...__calc_spacing_agent_gap_line({ p, axis, anchor }));
+
+        // lines for uniform gap loops (including self)
+
+        if (fwd !== -1) {
+          const loop = loops[fwd];
+          const gap = gaps[fwd];
+          const rects_loop = loop.map(
+            (idx) => anchors[aligned_anchors_idx[idx]]
+          );
+
+          lines.push(
+            __calc_spacing_loop_gap_line({
+              axis,
+              loop: rects_loop,
+              gap: gap,
+            })
+          );
+        }
+      });
+    }
+  }
+
+  // Only draw for whichever axis actually caused the snap
+  if (x) {
+    handle_axis({
+      ...x,
+      aligned_anchors_idx: x.aligned_anchors_idx,
+      anchors: main_anchors,
+      axis: "x",
+    });
+  }
+
+  if (y) {
+    handle_axis({
+      ...y,
+      aligned_anchors_idx: y.aligned_anchors_idx,
+      anchors: main_anchors,
+      axis: "y",
+    });
+  }
 
   return {
-    x: x_snap,
-    y: y_snap,
+    lines,
+  };
+}
+
+export type SnapGuide = {
+  points: cmath.Vector2[];
+  rules: cmath.ui.Rule[];
+  lines: cmath.ui.Line[];
+};
+
+export function guide(snapping: SnapToObjectsResult): SnapGuide {
+  const lines: cmath.ui.Line[] = [];
+  const points: cmath.Vector2[] = [];
+  const x_ray_offsets: number[] = [];
+  const y_ray_offsets: number[] = [];
+
+  // #region by_geometry
+  const by_geometry = __surface_snap_guide_by_geometry(snapping);
+
+  points.push(...by_geometry.points);
+  lines.push(...by_geometry.lines);
+  // #endregion by_geometry
+
+  // #region by_spacing
+  const by_spacing = __surface_snap_guide_by_spacing(snapping);
+  lines.push(...by_spacing.lines);
+  // #endregion by_spacing
+
+  const rays: cmath.ui.Rule[] = [
+    ...Array.from(new Set(x_ray_offsets)).map(
+      (offset) => ["x", offset] satisfies cmath.ui.Rule
+    ),
+    ...Array.from(new Set(y_ray_offsets)).map(
+      (offset) => ["y", offset] satisfies cmath.ui.Rule
+    ),
+  ];
+
+  return {
+    points,
+    rules: rays,
+    lines,
   };
 }
